requirements.txt
Django==4.2.10
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-ckeditor==6.7.3
python-dotenv==1.0.1
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-filter==24.1 